Cast up to unsigned longs, in case of small size_t's on non-64-bit machines
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@748 96dfbd92-d197-e211-9ca9-00110a534b34

+++ tests/main.c
@@ -124,7 +124,7 @@ int main(int argc, const char *argv[]) {
 		}
 	}
 	
-	printf("Compiled unit test results: %lu/%lu tests passed, %lu%% pass rate.\n", TOTAL_TESTS - rc, TOTAL_TESTS, 100 * (TOTAL_TESTS - rc) / TOTAL_TESTS);
+	printf("Compiled unit test results: %lu/%lu tests passed, %lu%% pass rate.\n", (unsigned long)TOTAL_TESTS - rc, (unsigned long)TOTAL_TESTS, 100UL * (TOTAL_TESTS - rc) / TOTAL_TESTS);
 	
 	return rc;
 }
